fixed weird seg faults and formatting
the system seg faults because of how some of the debug code works. i've fixed it so that it is robust enough to not seg fault. the issue was with the num_nw array that was hard coded to only work with 8 threads but gnumap still allos uses to use as many threads as you want. if you specify more than 8 threads, you can index out bounds and corrupt other variables, i noticed it corrupting the gILLUMINA and the ofstream of variables.